+++ src/apps/editor_sdl_main.zig
@@ -7,6 +7,7 @@ const swapchain_mod = engine.swapchain;
 const render = engine.renderer;
 const helper = engine.helper;
 const text = engine.textures;
+const input = engine.input;
 const c = engine.c;
 const print = std.debug.print;
 const sdl = engine.sdl;
@@ -14,6 +15,10 @@ const math = utils.math;
 const algo = utils.algo;
 const notify = utils.notify;
 const atlas_mod = utils.atlas;
+const lua_mod = engine.lua;
+//const lua = lua_mod.lua;
+const zlua = @import("zlua");
+const Lua = zlua.Lua;
 
 // Opaque = 0
 // Alpha = 1
@@ -24,7 +29,33 @@ const atlas_mod = utils.atlas;
 // 2. atlas_list is always sorted by id
 // 3. IDs are stable and never renumbered
 // 4. Editor never invents IDs
+fn MoveSprite(lua: *Lua) i32 {
+    const a = lua.toNumber(1) catch 0;
+    const b = lua.toNumber(2) catch 0;
+    lua.pushNumber(a + b);
+    return 1;
+}
 
+// ****************************************** CAMERA ************************************
+pub const Camera = struct {
+    pos: math.Vec2 = math.Vec2.ZERO,
+    view_proj: math.Mat4 = math.Mat4.IDENTITY,
+
+    pub fn Update(self: *Camera, screen_w: f32, screen_h: f32) void {
+
+        const proj = math.Ortho(0.0, screen_w, 0.0, screen_h);
+
+        const view = math.Mat4.TranslateWorld(proj , .{
+            .x = -self.pos.x,
+            .y = -self.pos.y,
+            .z = 0,
+        });
+
+        self.view_proj = view;
+
+    }
+
+};
 
 
 // ****************************************** ATLAS MANAGER **********************************
@@ -183,12 +214,32 @@ pub fn main() !void {
     // Window Creation
     var game_window = try sdl.Window.init(800, 600);
     defer game_window.deinit();
+
+    // Editor Input
+    var editor_input = input.EditorIntent{
+        .drag_speed = 0.05,
+    };
     
+    // Camera 
+    var camera = Camera{};
+   
     // Allocator
     var gpa = std.heap.GeneralPurposeAllocator(.{}){};
     defer _ = gpa.deinit();
     const allocator = gpa.allocator();
     
+    // Lua
+    var lua = try Lua.init(allocator);
+    defer lua.deinit();
+
+    lua.openLibs();
+
+    lua.pushInteger(42);
+    std.debug.print("{}\n", .{try lua.toInteger(1)});
+    try lua.doString("print('Lua is alive')");
+
+    lua.pushFunction(zlua.wrap(MoveSprite));
+    lua.setGlobal("move_sprite");
     // Core Creation
     var core = try core_mod.Core.init(true, allocator, &game_window);
     defer core.deinit(allocator);
@@ -198,7 +249,7 @@ pub fn main() !void {
     defer sc.deinit(&core, allocator, core.alloc_cb);
     
     // Renderer creation
-    var renderer = try render.Renderer.init(allocator, &core, &sc);
+    var renderer = try render.Renderer.init(allocator, &core, &sc, &game_window);
     defer renderer.deinit(allocator, &core);
 
     // Atlas Manager
@@ -235,21 +286,49 @@ pub fn main() !void {
         allocator.free(slot_uv.?.name);
     }
 
-    const slot_sprite_draw = helper.SpriteDraw{
+    const sprite_w: f32 = 400.0;
+    const sprite_h: f32 = 200.0;
+
+    const screen_w = @as(f32, @floatFromInt(game_window.screen_width));
+    const screen_h = @as(f32, @floatFromInt(game_window.screen_height));
+
+    const center_x: f32 = (screen_w - sprite_w) * 0.5;
+    const center_y: f32 = (screen_h - sprite_h) * 0.5;
+
+    var slot_sprite_draw = helper.SpriteDraw{
         .uv_min = slot_uv.?.uv_min,
         .uv_max = slot_uv.?.uv_max,
-        .sprite_pos   = .{ 0, 0},        // world position (your choice)
-        .sprite_scale = .{ 0, 0 },// world size (or whatever units you use)
+        .sprite_pos   = .{ center_x, center_y},        // world position (your choice)
+        .sprite_scale = .{ 300.0 , 200.0 },// world size (or whatever units you use)
         .sprite_rotation = .{1.0, 0.0}, // cos=1, sin=0 (no rotation)
         .tint = .{ 1, 1, 1, 1 },   // no tint
         .atlas_id = 0,
     };
 
+    lua.pushLightUserdata(&slot_sprite_draw);
+    lua.setGlobal("test");
+
+    const slot_sprite_draw_s = helper.SpriteDraw{
+        .uv_min = slot_uv.?.uv_min,
+        .uv_max = slot_uv.?.uv_max,
+        .sprite_pos   = .{ center_x + 300, center_y},        // world position (your choice)
+        .sprite_scale = .{ 300.0 , 200.0 },// world size (or whatever units you use)
+        .sprite_rotation = .{1.0, 0.0}, // cos=1, sin=0 (no rotation)
+        .tint = .{ 1, 1, 1, 1 },   // no tint
+        .atlas_id = 0,
+    };
     try sprite_draws.append(allocator, slot_sprite_draw);
+    try sprite_draws.append(allocator, slot_sprite_draw_s);
 
     while (!game_window.should_close){
-
+        
         game_window.pollEvents(&renderer);
+        input.BuildEditorIntent(&editor_input, game_window.raw_input);
+        camera.pos = math.Vec2.Add(camera.pos , editor_input.drag_delta);
+        camera.Update(
+            @floatFromInt(game_window.screen_width), 
+            @floatFromInt(game_window.screen_height)
+        );
         _ = try atlas_notifier.poll();
 
         if (atlas_manager.metadata_dirty){
@@ -273,7 +352,9 @@ pub fn main() !void {
             &sc, 
             &game_window, 
             allocator, 
-            sprite_draws.items);
+            sprite_draws.items,
+            camera.view_proj,
+        );
 
     }