+++ src/engine/engine.zig
@@ -5,4 +5,5 @@ pub const renderer = @import("render.zig");
 pub const helper = @import("helper.zig");
 pub const textures = @import("textures.zig");
 pub const sdl = @import("sdl.zig");
+pub const input = @import("controls.zig");